Moved contract clients, hooks, and selectors out of proposal module adapters.
