+++ packages/state/contracts/index.ts
@@ -9,6 +9,10 @@ export {
   CwPayrollFactoryQueryClient,
 } from './CwPayrollFactory'
 export {
+  CwProposalSingleV1Client,
+  CwProposalSingleV1QueryClient,
+} from './CwProposalSingle.v1'
+export {
   CwTokenfactoryIssuerClient,
   CwTokenfactoryIssuerQueryClient,
 } from './CwTokenfactoryIssuer'
@@ -23,6 +27,22 @@ export {
   DaoPreProposeApproverClient,
   DaoPreProposeApproverQueryClient,
 } from './DaoPreProposeApprover'
+export {
+  DaoPreProposeMultipleClient,
+  DaoPreProposeMultipleQueryClient,
+} from './DaoPreProposeMultiple'
+export {
+  DaoPreProposeSingleClient,
+  DaoPreProposeSingleQueryClient,
+} from './DaoPreProposeSingle'
+export {
+  DaoProposalMultipleClient,
+  DaoProposalMultipleQueryClient,
+} from './DaoProposalMultiple'
+export {
+  DaoProposalSingleV2Client,
+  DaoProposalSingleV2QueryClient,
+} from './DaoProposalSingle.v2'
 export { DaoVotingCw4QueryClient } from './DaoVotingCw4'
 export {
   DaoVotingCw20StakedClient,
