+++ packages/state/recoil/selectors/contracts/index.ts
@@ -5,12 +5,18 @@ export * as Cw20StakeSelectors from './Cw20Stake'
 export * as Cw721BaseSelectors from './Cw721Base'
 export * as CwAdminFactorySelectors from './CwAdminFactory'
 export * as CwPayrollFactorySelectors from './CwPayrollFactory'
+export * as CwProposalSingleV1Selectors from './CwProposalSingle.v1'
 export * as CwTokenfactoryIssuerSelectors from './CwTokenfactoryIssuer'
 export * as CwTokenSwapSelectors from './CwTokenSwap'
 export * as CwVestingSelectors from './CwVesting'
 export * as DaoCoreV2Selectors from './DaoCore.v2'
 export * as DaoPreProposeApprovalSingleSelectors from './DaoPreProposeApprovalSingle'
 export * as DaoPreProposeApproverSelectors from './DaoPreProposeApprover'
+export * as DaoPreProposeMultipleSelectors from './DaoPreProposeMultiple'
+export * as DaoPreProposeSingleSelectors from './DaoPreProposeSingle'
+export * as DaoProposalMultipleSelectors from './DaoProposalMultiple'
+export * as DaoProposalSingleCommonSelectors from './DaoProposalSingle.common'
+export * as DaoProposalSingleV2Selectors from './DaoProposalSingle.v2'
 export * as DaoVotingCw4Selectors from './DaoVotingCw4'
 export * as DaoVotingCw20StakedSelectors from './DaoVotingCw20Staked'
 export * as DaoVotingCw721StakedSelectors from './DaoVotingCw721Staked'
