+++ packages/stateful/hooks/contracts/CwProposalSingle.v1.ts
@@ -4,15 +4,15 @@ import { ExecuteResult } from '@cosmjs/cosmwasm-stargate'
 import { useCallback } from 'react'
 import { useRecoilValueLoadable } from 'recoil'
 
-import { useChain } from '@dao-dao/stateless'
-import { FunctionKeyOf } from '@dao-dao/types'
-
-import { useSyncWalletSigner } from '../../../../hooks'
-import { CwProposalSingleV1Client as ExecuteClient } from './CwProposalSingle.v1.client'
+import { CwProposalSingleV1Client as ExecuteClient } from '@dao-dao/state/contracts/CwProposalSingle.v1'
 import {
   ExecuteClientParams,
   executeClient,
-} from './CwProposalSingle.v1.recoil'
+} from '@dao-dao/state/recoil/selectors/contracts/CwProposalSingle.v1'
+import { useChain } from '@dao-dao/stateless'
+import { FunctionKeyOf } from '@dao-dao/types'
+
+import { useSyncWalletSigner } from '../useSyncWalletSigner'
 
 // This hook wrapper lets us easily make hooks out of all execution functions on
 // the contract clients, without having to fetch the `executeClient` selector as
