+++ packages/stateful/hooks/contracts/DaoProposalMultiple.ts
@@ -4,15 +4,15 @@ import { ExecuteResult } from '@cosmjs/cosmwasm-stargate'
 import { useCallback } from 'react'
 import { useRecoilValueLoadable } from 'recoil'
 
-import { useChain } from '@dao-dao/stateless'
-import { FunctionKeyOf } from '@dao-dao/types'
-
-import { useSyncWalletSigner } from '../../../../hooks'
-import { DaoProposalMultipleClient as ExecuteClient } from './DaoProposalMultiple.client'
+import { DaoProposalMultipleClient as ExecuteClient } from '@dao-dao/state/contracts/DaoProposalMultiple'
 import {
   ExecuteClientParams,
   executeClient,
-} from './DaoProposalMultiple.recoil'
+} from '@dao-dao/state/recoil/selectors/contracts/DaoProposalMultiple'
+import { useChain } from '@dao-dao/stateless'
+import { FunctionKeyOf } from '@dao-dao/types'
+
+import { useSyncWalletSigner } from '../useSyncWalletSigner'
 
 // This hook wrapper lets us easily make hooks out of all execution functions on
 // the contract clients, without having to fetch the `executeClient` selector as
