+++ packages/stateful/hooks/contracts/DaoProposalSingle.v2.ts
@@ -4,15 +4,15 @@ import { ExecuteResult } from '@cosmjs/cosmwasm-stargate'
 import { useCallback } from 'react'
 import { useRecoilValueLoadable } from 'recoil'
 
-import { useChain } from '@dao-dao/stateless'
-import { FunctionKeyOf } from '@dao-dao/types'
-
-import { useSyncWalletSigner } from '../../../../hooks'
-import { DaoProposalSingleV2Client as ExecuteClient } from './DaoProposalSingle.v2.client'
+import { DaoProposalSingleV2Client as ExecuteClient } from '@dao-dao/state/contracts/DaoProposalSingle.v2'
 import {
   ExecuteClientParams,
   executeClient,
-} from './DaoProposalSingle.v2.recoil'
+} from '@dao-dao/state/recoil/selectors/contracts/DaoProposalSingle.v2'
+import { useChain } from '@dao-dao/stateless'
+import { FunctionKeyOf } from '@dao-dao/types'
+
+import { useSyncWalletSigner } from '../useSyncWalletSigner'
 
 // This hook wrapper lets us easily make hooks out of all execution functions on
 // the contract clients, without having to fetch the `executeClient` selector as
