+++ packages/stateful/hooks/contracts/index.ts
@@ -2,7 +2,12 @@ export * as Cw20BaseHooks from './Cw20Base'
 export * as Cw20StakeHooks from './Cw20Stake'
 export * as Cw721BaseHooks from './Cw721Base'
 export * as CwAdminFactoryHooks from './CwAdminFactory'
+export * as CwProposalSingleV1Hooks from './CwProposalSingle.v1'
 export * as CwVesting from './CwVesting'
+export * as DaoPreProposeMultipleHooks from './DaoPreProposeMultiple'
+export * as DaoPreProposeSingleHooks from './DaoPreProposeSingle'
+export * as DaoProposalMultipleHooks from './DaoProposalMultiple'
+export * as DaoProposalSingleV2Hooks from './DaoProposalSingle.v2'
 export * as DaoVotingCw721StakedHooks from './DaoVotingCw721Staked'
 export * as DaoVotingNativeStakedHooks from './DaoVotingNativeStaked'
 export * as DaoVotingTokenStakedHooks from './DaoVotingTokenStaked'
