+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalMultiple/common/actions/UpdatePreProposeConfig/index.tsx
@@ -3,7 +3,11 @@ import { useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import { constSelector, useRecoilValueLoadable } from 'recoil'
 
-import { genericTokenSelector, isContractSelector } from '@dao-dao/state'
+import {
+  DaoPreProposeMultipleSelectors,
+  genericTokenSelector,
+  isContractSelector,
+} from '@dao-dao/state'
 import { GearEmoji, useCachedLoadingWithError } from '@dao-dao/stateless'
 import {
   ActionComponent,
@@ -33,7 +37,6 @@ import {
 import { useActionOptions } from '../../../../../../actions'
 import { useVotingModuleAdapter } from '../../../../../../voting-module-adapter'
 import { PRE_PROPOSE_CONTRACT_NAMES } from '../../../constants'
-import { configSelector } from '../../../contracts/DaoPreProposeMultiple.recoil'
 import {
   UpdatePreProposeConfigComponent,
   UpdatePreProposeConfigData,
@@ -134,7 +137,7 @@ export const makeUpdatePreProposeConfigActionMaker =
         useCommonGovernanceTokenInfo?.() ?? {}
 
       const configLoading = useCachedLoadingWithError(
-        configSelector({
+        DaoPreProposeMultipleSelectors.configSelector({
           chainId,
           contractAddress: preProposeAddress,
           params: [],
