+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalMultiple/common/actions/UpdateProposalConfig/index.tsx
@@ -1,7 +1,10 @@
 import { useCallback } from 'react'
 import { constSelector } from 'recoil'
 
-import { isContractSelector } from '@dao-dao/state/recoil'
+import {
+  DaoProposalMultipleSelectors,
+  isContractSelector,
+} from '@dao-dao/state/recoil'
 import {
   BallotDepositEmoji,
   useCachedLoadingWithError,
@@ -23,7 +26,6 @@ import {
 import { makeWasmMessage, objectMatchesStructure } from '@dao-dao/utils'
 
 import { CONTRACT_NAMES } from '../../../constants'
-import { configSelector } from '../../../contracts/DaoProposalMultiple.recoil'
 import { UpdateProposalConfigComponent as Component } from './UpdateProposalConfigComponent'
 
 export interface UpdateProposalConfigData {
@@ -100,7 +102,7 @@ export const makeUpdateProposalConfigActionMaker =
   ({ t, context, chain: { chain_id: chainId } }) => {
     const useDefaults: UseDefaults<UpdateProposalConfigData> = () => {
       const proposalModuleConfig = useCachedLoadingWithError(
-        configSelector({
+        DaoProposalMultipleSelectors.configSelector({
           chainId,
           contractAddress: proposalModuleAddress,
         })
@@ -135,7 +137,7 @@ export const makeUpdateProposalConfigActionMaker =
       UpdateProposalConfigData
     > = () => {
       const proposalModuleConfig = useCachedLoadingWithError(
-        configSelector({
+        DaoProposalMultipleSelectors.configSelector({
           chainId,
           contractAddress: proposalModuleAddress,
         })
