+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/common/actions/UpdateProposalConfigV2/index.tsx
@@ -1,7 +1,10 @@
 import { useCallback } from 'react'
 import { constSelector } from 'recoil'
 
-import { isContractSelector } from '@dao-dao/state/recoil'
+import {
+  DaoProposalSingleV2Selectors,
+  isContractSelector,
+} from '@dao-dao/state/recoil'
 import {
   BallotDepositEmoji,
   useCachedLoadingWithError,
@@ -25,7 +28,6 @@ import {
 } from '@dao-dao/utils'
 
 import { CONTRACT_NAMES } from '../../../constants'
-import { configSelector } from '../../../contracts/DaoProposalSingle.v2.recoil'
 import { UpdateProposalConfigComponent as Component } from './UpdateProposalConfigComponent'
 
 export interface UpdateProposalConfigData {
@@ -139,7 +141,7 @@ export const makeUpdateProposalConfigV2ActionMaker =
 
     const useDefaults: UseDefaults<UpdateProposalConfigData> = () => {
       const proposalModuleConfig = useCachedLoadingWithError(
-        configSelector({
+        DaoProposalSingleV2Selectors.configSelector({
           chainId,
           contractAddress: proposalModuleAddress,
         })
@@ -173,7 +175,7 @@ export const makeUpdateProposalConfigV2ActionMaker =
       UpdateProposalConfigData
     > = () => {
       const proposalModuleConfig = useCachedLoadingWithError(
-        configSelector({
+        DaoProposalSingleV2Selectors.configSelector({
           chainId,
           contractAddress: proposalModuleAddress,
         })
