+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/common/components/NewProposal.tsx
@@ -5,7 +5,11 @@ import toast from 'react-hot-toast'
 import { useTranslation } from 'react-i18next'
 import { useRecoilCallback, useRecoilValueLoadable } from 'recoil'
 
-import { DaoCoreV2Selectors, blocksPerYearSelector } from '@dao-dao/state'
+import {
+  DaoCoreV2Selectors,
+  DaoProposalSingleCommonSelectors,
+  blocksPerYearSelector,
+} from '@dao-dao/state'
 import {
   NewProposalTitleDescriptionHeader,
   NewProposal as StatelessNewProposal,
@@ -27,7 +31,6 @@ import {
 
 import { useLoadedActionsAndCategories } from '../../../../../actions'
 import { useMembership, useWallet } from '../../../../../hooks'
-import { configSelector } from '../../contracts/DaoProposalSingle.common.recoil'
 import { makeGetProposalInfo } from '../../functions'
 import {
   NewProposalData,
@@ -153,7 +156,7 @@ export const NewProposal = ({
             )
 
           const config = await snapshot.getPromise(
-            configSelector({
+            DaoProposalSingleCommonSelectors.configSelector({
               chainId,
               contractAddress: options.proposalModule.address,
             })
