+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/hooks/useCastVote.ts
@@ -5,10 +5,12 @@ import { ContractVersion } from '@dao-dao/types'
 import { Vote } from '@dao-dao/types/contracts/DaoProposalSingle.common'
 import { processError } from '@dao-dao/utils'
 
+import {
+  CwProposalSingleV1Hooks,
+  DaoProposalSingleV2Hooks,
+} from '../../../../hooks'
 import { useWallet } from '../../../../hooks/useWallet'
 import { useProposalModuleAdapterOptions } from '../../../react'
-import { useVote as useVoteV1 } from '../contracts/CwProposalSingle.v1.hooks'
-import { useVote as useVoteV2 } from '../contracts/DaoProposalSingle.v2.hooks'
 import { useLoadingWalletVoteInfo } from './useLoadingWalletVoteInfo'
 
 export const useCastVote = (onSuccess?: () => void | Promise<void>) => {
@@ -16,7 +18,9 @@ export const useCastVote = (onSuccess?: () => void | Promise<void>) => {
   const { isWalletConnected, address: walletAddress = '' } = useWallet()
 
   const _castVote = (
-    proposalModule.version === ContractVersion.V1 ? useVoteV1 : useVoteV2
+    proposalModule.version === ContractVersion.V1
+      ? CwProposalSingleV1Hooks.useVote
+      : DaoProposalSingleV2Hooks.useVote
   )({
     contractAddress: proposalModule.address,
     sender: walletAddress ?? '',
