+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/hooks/useLoadingDepositInfo.ts
@@ -1,5 +1,9 @@
 import { constSelector } from 'recoil'
 
+import {
+  CwProposalSingleV1Selectors,
+  DaoPreProposeSingleSelectors,
+} from '@dao-dao/state'
 import { useCachedLoadable } from '@dao-dao/stateless'
 import {
   CheckedDepositInfo,
@@ -12,8 +16,6 @@ import { ProposalResponse as ProposalV1Response } from '@dao-dao/types/contracts
 import { DepositInfoResponse as DepositInfoPreProposeResponse } from '@dao-dao/types/contracts/DaoPreProposeSingle'
 
 import { useProposalModuleAdapterOptions } from '../../../react/context'
-import { proposalSelector as proposalV1Selector } from '../contracts/CwProposalSingle.v1.recoil'
-import { depositInfoSelector as depositInfoV2Selector } from '../contracts/DaoPreProposeSingle.recoil'
 
 export const useLoadingDepositInfo = (): LoadingData<
   CheckedDepositInfo | undefined
@@ -29,7 +31,7 @@ export const useLoadingDepositInfo = (): LoadingData<
   >(
     // V1 does not support pre-propose.
     version === ContractVersion.V1
-      ? proposalV1Selector({
+      ? CwProposalSingleV1Selectors.proposalSelector({
           chainId,
           contractAddress: address,
           params: [
@@ -42,7 +44,7 @@ export const useLoadingDepositInfo = (): LoadingData<
       prePropose &&
         // Approver does not have deposit info.
         prePropose.type !== PreProposeModuleType.Approver
-      ? depositInfoV2Selector({
+      ? DaoPreProposeSingleSelectors.depositInfoSelector({
           chainId,
           contractAddress: prePropose.address,
           params: [
